+++ tests/PhpWord/Reader/Word2007/StyleTest.php
@@ -19,6 +19,7 @@ namespace PhpOffice\PhpWord\Reader\Word2007;
 
 use PhpOffice\PhpWord\AbstractTestReader;
 use PhpOffice\PhpWord\SimpleType\TblWidth;
+use PhpOffice\PhpWord\SimpleType\VerticalJc;
 use PhpOffice\PhpWord\Style;
 use PhpOffice\PhpWord\Style\Table;
 use PhpOffice\PhpWord\Style\TablePosition;
@@ -212,5 +213,17 @@ class StyleTest extends AbstractTestReader
 
         $this->getDocumentFromString(array('styles' => $documentXml));
         $this->assertInstanceOf('PhpOffice\\PhpWord\\Style\\Font', Style::getStyle($name));
+    }
+
+    public function testPageVerticalAlign()
+    {
+        $documentXml = '<w:sectPr>
+            <w:vAlign w:val="center"/>
+        </w:sectPr>';
+
+        $phpWord = $this->getDocumentFromString(array('document' => $documentXml));
+
+        $sectionStyle = $phpWord->getSection(0)->getStyle();
+        $this->assertEquals(VerticalJc::CENTER, $sectionStyle->getVAlign());
     }
 }